Extracting parseNamed out to be tested.
This commit extracts the function parsedName along with the regex for the options into is own JavaScript file. This way it can be tested. I
included the parsedName file in server.js and it was still functioning properly. This commit also writes unit tests for all the different
parameters for file names and making sure the opts for render-sender are being set correctly. Possibly a few more unit tests with multiple
parameters could be tested just to make sure everything is working properly.